requirements.txt
fastapi
uvicorn[standard]
websockets
PyYAML
edge-tts
httpx
sherpa-onnx
faster-whisper
openai
pysbd
langdetect
chardet
loguru
onnxruntime
requests
tqdm
soundfile
numpy
google-generativeai